1. Update XiaoDB, WriteBTreeNode and BTree head to disk file with UTs. 2. TODO - UT mentioned above not ALL PASS.

+++ XiaoDB/BTreeCommon.h
@@ -30,7 +30,7 @@ typedef uint64 PointerType;
 struct Key
 {
     uchar *mKey;
-    uint8 mKeyLen;
+    uint32 mKeyLen;
 
     /**
      * Ctor
